requirements.txt
nltk
sentence-transformers
faiss-cpu
PyMuPDF
transformers
torch